animasi catatan

+++ src/script/view/home.js
@@ -1,5 +1,6 @@
-import { showLoading, hideLoading } from "../utils.js";
+import { showLoading, hideLoading, animeToNote, animeOutFromNote } from "../utils.js";
 const Swal = require("sweetalert2");
+const anime = require('animejs').default
 
 const baseUrl = "https://notes-api.dicoding.dev/v2";
 
@@ -81,6 +82,11 @@ const insertNote = async (note) => {
 };
 
 const archiveNote = async (noteId) => {
+  const noteItem = document.querySelector(`[data-noteid="${noteId}"]`);
+  if (!noteItem) return; 
+
+  await animeOutFromNote(noteItem)
+
   try {
     const options = {
       method: "POST",
@@ -102,6 +108,11 @@ const archiveNote = async (noteId) => {
 };
 
 const unarchiveNote = async (noteId) => {
+  const noteItem = document.querySelector(`[data-noteid="${noteId}"]`);
+  if (!noteItem) return; 
+
+  await animeOutFromNote(noteItem)
+
   try {
     const options = {
       method: "POST",
@@ -126,6 +137,8 @@ const unarchiveNote = async (noteId) => {
 };
 
 const removeNote = async (noteId) => {
+  const noteItem = document.querySelector(`[data-noteid="${noteId}"]`);
+  if (!noteItem) return; 
   try {
     const options = {
       method: "DELETE",
@@ -136,6 +149,8 @@ const removeNote = async (noteId) => {
 
     const response = await fetch(`${baseUrl}/notes/${noteId}`, options);
     const data = await response.json();
+
+    await animeOutFromNote(noteItem)
 
     showResponseMessage(data.message, "success");
     await getNote();
@@ -180,6 +195,8 @@ const alertConfirm = async (
   return result.isConfirmed;
 };
 
+
+
 const displayNotes = (notes) => {
   const containerElement = document.querySelector("#noteListContainer");
   const listElement = containerElement.querySelector(".list");
@@ -193,14 +210,18 @@ const displayNotes = (notes) => {
     const buttonContainer = noteItemElement.querySelector(".noteitem-button");
 
     const archiveButton = document.createElement("button");
-    archiveButton.textContent = "Arsipkan";
-    archiveButton.classList.add("button-arsip");
+    const spanIcon = document.createElement('span')
+    spanIcon.classList.add("mdi--archive-arrow-down")
+    archiveButton.title = "Arsipkan";
     archiveButton.addEventListener("click", () => archiveNote(note.id));
     buttonContainer.appendChild(archiveButton);
+    archiveButton.appendChild(spanIcon)
 
     const deleteButton = document.createElement("button");
-    deleteButton.textContent = "Hapus";
-    deleteButton.classList.add("button-delete");
+    const spanIconDelete = document.createElement('span')
+    spanIconDelete.classList.add("mdi--delete")
+    deleteButton.title = "Hapus";
+    // deleteButton.classList.add("button-delete");
     deleteButton.addEventListener("click", async () => {
       const isConfirmed = await alertConfirm(
         "Yakin ingin menghapus catatan ini?",
@@ -210,8 +231,12 @@ const displayNotes = (notes) => {
       if (isConfirmed) removeNote(note.id);
     });
     buttonContainer.appendChild(deleteButton);
+    deleteButton.appendChild(spanIconDelete)
 
     listElement.appendChild(noteItemElement);
+
+    animeToNote(noteItemElement)
+
   });
 };
 
@@ -228,14 +253,18 @@ const displayArsipNotes = (notes) => {
     const buttonContainer = noteItemElement.querySelector(".noteitem-button");
 
     const archiveButton = document.createElement("button");
-    archiveButton.textContent = "Batalkan Arsip";
-    archiveButton.classList.add("button-arsip");
+    const spanIcon = document.createElement('span')
+    spanIcon.classList.add("mdi--archive-arrow-up")
+    archiveButton.title = "Batalkan Arsip";
     archiveButton.addEventListener("click", () => unarchiveNote(note.id));
     buttonContainer.appendChild(archiveButton);
+    archiveButton.appendChild(spanIcon)
 
     const deleteButton = document.createElement("button");
-    deleteButton.textContent = "Hapus";
-    deleteButton.classList.add("button-delete");
+    const spanIconDelete = document.createElement('span')
+    spanIconDelete.classList.add("mdi--delete")
+    deleteButton.title = "Hapus";
+    deleteButton.title = "Hapus";
     deleteButton.addEventListener("click", async () => {
       const isConfirmed = await alertConfirm(
         "Yakin ingin menghapus catatan ini?",
@@ -245,8 +274,10 @@ const displayArsipNotes = (notes) => {
       if (isConfirmed) removeNote(note.id);
     });
     buttonContainer.appendChild(deleteButton);
+    deleteButton.appendChild(spanIconDelete)
 
     listElement.appendChild(noteItemElement);
+    animeToNote(noteItemElement)
   });
 };
